Move IShellItemArray into shobjidl_core, leaving a reexport.
